SERVER/requirements.txt
blinker==1.7.0
certifi==2023.11.17
cftime==1.6.3
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
geographiclib==2.0
geopy==2.4.1
HeapDict==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
netCDF4==1.6.5
numpy==1.26.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
